Combine subjobs and state_updates
Both deal with job updates, makes subjobs easier to mock out in tests

Pass state updates into KBaseRunner for easier unit testing in the
future
